feat(embed): polish embed service and migrate pgvector store to main library
Update uvicorn server to support PostgreSQL/pgvector backend, centralize DB schemas, move pgvector store logic to the shared cryptotrading library, and add unit tests.
