+++ .agent/core/activeContext.md
@@ -1,16 +1,33 @@
-# Active Context: Rework Order Book Snapshot Panel
+# Active Context: Polish Embed Service & Move pgvector Store
 
 ## Quick Reference
-- **Feature**: Rework Order Book snapshot panel into real-time depth visualization
-- **Branch**: `feature/order-book-rework`
+- **Feature**: Polish Embed Service & Move pgvector Store to main library
+- **Branch**: `feature/polish-embed-service`
+- **Plan File**: `.agent/plans/move-pgvector-store-plan.md`
 - **Status**: Completed & Verified ✅
 
 ## Executive Summary
-Successfully redesigned the OrderBookPanel component to provide maximum price direction utility. The new visualization features a real-time cumulative depth wall chart built with ECharts (including a vertical midpoint marker), a bidirectional buy/sell pressure meter that alerts on bullish/bearish imbalances, and whale limit wall callouts displaying major support and resistance order sizes. The Vite production build compiles successfully.
+Successfully migrated the pgvector database adapter logic out of `services/embed` and into the main `cryptotrading` package under `src/cryptotrading/data/pgvector_store.py`. Integrated the `trade_setups` table and HNSW index creation centrally in `postgres.py`. Wrote and verified comprehensive unit tests.
 
-## Key Files to Create/Modify
-- [OrderBookPanel.jsx](file:///home/miles/Development/notebooks/CryptoTrading/frontend/src/components/OrderBookPanel.jsx): Reworked with depth walls chart, pressure meter, and whale alerts.
+## Architecture Overview
+The pgvector database model for trade setups is centralized. The table schemas (including HNSW indices) are integrated into `postgres.py`'s centralized initialization. `TradeEmbeddingDB` links directly to the shared `postgres.py` connection pool.
 
+## Tech Stack for This Feature
+- **FastAPI**: Endpoint handler
+- **PostgreSQL + pgvector**: Vector and metadata store
+- **asyncpg**: Async database connector
+- **PyTorch**: Contrastive representation encoder
 
+## Key Files Created/Modified
+- [src/cryptotrading/data/postgres.py](file:///home/miles/Development/notebooks/CryptoTrading/src/cryptotrading/data/postgres.py): Added `trade_setups` schema and index initialization centrally.
+- [src/cryptotrading/data/pgvector_store.py](file:///home/miles/Development/notebooks/CryptoTrading/src/cryptotrading/data/pgvector_store.py): Centrally defined pgvector database adapter.
+- [services/embed/database/pgvector_store.py](file:///home/miles/Development/notebooks/CryptoTrading/services/embed/database/pgvector_store.py): Redirected imports to the main package.
+- [tests/test_pgvector_store.py](file:///home/miles/Development/notebooks/CryptoTrading/tests/test_pgvector_store.py): Unit tests for pgvector_store and compatibility.
 
-
+## Verification & Validation
+- Verified using `docker compose up -d --build embed` to compile the Docker image and spin up the service.
+- Checked container logs to ensure database connection and schema initialization completed without errors.
+- Verified `/health` endpoint responds with a successful status.
+- Executed unit tests in virtual environment:
+  - `src/.venv/bin/pytest tests/test_embed_service.py` -> **Passed** (1/1)
+  - `src/.venv/bin/pytest tests/test_pgvector_store.py` -> **Passed** (7/7)
