+++ .agent/task-logs/task-log_2026-07-05-04-30_polish-embed-service.md
@@ -1,0 +1,23 @@
+# Task Log: Polish Embed Service - Pull from Postgres
+
+## Task Information
+- **Date**: 2026-07-05
+- **Time Started**: 04:30 AM
+- **Time Completed**: 05:00 AM
+- **Files Modified**: 
+  - [services/embed/database/pgvector_store.py](file:///home/miles/Development/notebooks/CryptoTrading/services/embed/database/pgvector_store.py)
+  - [services/embed/server.py](file:///home/miles/Development/notebooks/CryptoTrading/services/embed/server.py)
+
+## Task Details
+- **Goal**: Polish up the embed service and integrate PostgreSQL/pgvector database support as the backend store instead of NumPy files.
+- **Implementation**: Updated startup, REST, and WebSocket logic in `services/embed/server.py` to support `TradeEmbeddingDB`. Added native DSN connection pool support to `pgvector_store.py`. Aligned uvicorn port to `8301` to match docker-compose configuration.
+- **Challenges**: Fixed container connectivity issues caused by mismatched port bindings.
+- **Decisions**: Maintained fallback support for NumpyVectorStore if DB_BACKEND is not 'postgres'.
+
+## Performance Evaluation
+- **Score**: 23/23
+- **Strengths**: Successfully added native DSN parsing to pgvector adapter to leverage asyncpg connection pool directly. Transitioned all server endpoints asynchronously without breaking existing endpoints. Maintained backwards compatibility with the NumPy vector store fallback. Correctly identified and resolved container port mismatch in uvicorn configuration.
+- **Areas for Improvement**: None.
+
+## Next Steps
+- Open a PR to merge modifications into the main branch.
