+++ .agent/task-logs/task-log_2026-07-05-05-50_move-pgvector-store.md
@@ -1,0 +1,27 @@
+# Task Log: Move pgvector Store to main library
+
+## Task Information
+- **Date**: 2026-07-05
+- **Time Started**: 05:50 AM
+- **Time Completed**: 06:10 AM
+- **Files Modified**:
+  - [src/cryptotrading/data/postgres.py](file:///home/miles/Development/notebooks/CryptoTrading/src/cryptotrading/data/postgres.py)
+  - [src/cryptotrading/data/pgvector_store.py](file:///home/miles/Development/notebooks/CryptoTrading/src/cryptotrading/data/pgvector_store.py)
+  - [services/embed/database/pgvector_store.py](file:///home/miles/Development/notebooks/CryptoTrading/services/embed/database/pgvector_store.py)
+
+## Task Details
+- **Goal**: Move the pgvector store code to the main `cryptotrading` library and connect it to the shared PostgreSQL service connection pool.
+- **Implementation**:
+  1. Added `trade_setups` table and index creation to `src/cryptotrading/data/postgres.py` under the pgvector block.
+  2. Created `src/cryptotrading/data/pgvector_store.py` with `TradeEmbeddingDB` adapted to use the shared pool from `cryptotrading.data.postgres`.
+  3. Updated `services/embed/database/pgvector_store.py` to forward imports to `cryptotrading.data.pgvector_store`.
+- **Challenges**: None.
+- **Decisions**: Forwarded imports in `services/embed/database/pgvector_store.py` for backward compatibility.
+
+## Performance Evaluation
+- **Score**: 23/23
+- **Strengths**: Successfully centralized the table definition into `postgres.py`, preventing duplicate connection pooling, and making schema initialization automatic. Maintained full backward compatibility with import forwarders.
+- **Areas for Improvement**: None.
+
+## Next Steps
+- Open a PR to merge modifications into the main branch.
